fix bug introduced in 8f65309
Since Z-H-Sun/ADAblock@8f65309, there is a bug in saving 'image_figure_4_b.png', 'image_figure_4_c.png', and 'image_regions_defects_NEW.png' for some versions of ImageJ: Although the pictures can show correctly in the ImageJ window, but cannot display correct colors when saving. This might be because overlaying RGB images onto a channel stack by `Add Image` is not supported. Fixed in this commit. Also, saving a TIFF copy for each of the above files does not make sense here, as the layer information is minimal; deleted.